media-tool/requirements.txt
PyYAML
Pillow
pillow_heif
piexif
pytest
